move group_by_reducer serialization tests to kernel tests
